+++ app/src/main/AndroidManifest.xml
@@ -15,22 +15,22 @@
         android:theme="@style/AppTheme"
         tools:targetApi="31">
         <activity
-            android:name=".AddScrapActivity"
+            android:name=".ui.AddScrapActivity"
             android:exported="false" />
         <activity
-            android:name=".ScrapDetailEditMemoActivity"
+            android:name=".ui.ScrapDetailEditMemoActivity"
             android:exported="false" />
         <activity
-            android:name=".ScrapDetailActivity"
+            android:name=".ui.ScrapDetailActivity"
             android:exported="false" />
         <activity
-            android:name=".AddCategoryActivity"
+            android:name=".ui.AddCategoryActivity"
             android:exported="false" />
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.MainActivity"
             android:exported="false" />
         <activity
-            android:name=".LoginActivity"
+            android:name=".ui.LoginActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
