+++ app/src/main/java/com/scrap/scrap2024/ui/AddCategoryActivity.kt
@@ -1,4 +1,4 @@
-package com.scrap.scrap2024
+package com.scrap.scrap2024.ui
 
 import android.app.Activity
 import android.content.Intent
@@ -12,6 +12,7 @@ import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
+import com.scrap.scrap2024.R
 import com.scrap.scrap2024.databinding.ActivityAddCategoryBinding
