+++ app/src/main/res/layout/activity_scrap_detail_edit_memo.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ScrapDetailEditMemoActivity">
+    tools:context=".ui.ScrapDetailEditMemoActivity">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/constraintLayout"
