+++ app/src/main/res/layout/fragment_category.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/background"
-    tools:context=".nav.CategoryFragment">
+    tools:context=".ui.nav.CategoryFragment">
 
     <LinearLayout
         android:id="@+id/linearLayout"
