+++ app/src/main/res/layout/fragment_favorite.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/background"
-    tools:context=".nav.FavoriteFragment">
+    tools:context=".ui.nav.FavoriteFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
