+++ app/src/main/res/layout/fragment_mypage.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/background"
-    tools:context=".nav.MypageFragment">
+    tools:context=".ui.nav.MypageFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
