+++ app/src/main/res/layout/fragment_scrap.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/background"
-    tools:context=".nav.ScrapFragment">
+    tools:context=".ui.nav.ScrapFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/constraintLayout"
